Enabled tag for data columns
Columns can now be individually checked as data columns. Columns marked as data won't be enclosed in double quotes, everything else will be. This concept also remembers the columns that were checked as data and the ones that were not. Need to perform the other changes for Prefix and Suffix