Re-enabled column ordering
Columns can be re-ordered using drag-and-drop. Any changes made to the columns will transition even if the column is moved.
Switched array lists to HashMaps for: Prefix, Suffix, Find, Replace. New entries are added to the HashMap when the "update field" button is clicked. The entry is created in the HashMap as:
Key: Column header + |<Field>. For example: Layer|Prefix
There may be room for improvement in the future but it works for now.